requirements.txt
Django==2.2
djangorestframework==3.9.0
Pillow==9.2.0
